# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: Agentic Control Plane API
  version: 0.2.0
  summary: Policy, approval, execution, and audit APIs for governed agent actions
  description: |
    v0.2 extends the base control-plane lifecycle with richer, domain-specific schemas for:
      - Git and pull request actions
      - Deploy and cloud actions
      - Email and calendar actions
      - Payment and procurement actions

    Core lifecycle:
      intent -> policy decision -> approval (optional) -> execution grant -> execution -> trace
servers:
  - url: https://api.withaileron.ai
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /v1/enterprises/me:
    get:
      tags: [Enterprises]
      summary: Get current enterprise
      description: Returns the enterprise for the authenticated user.
      operationId: getCurrentEnterprise
      responses:
        "200":
          description: Current enterprise
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enterprise"
        "401":
          $ref: "#/components/responses/Unauthorized"

    patch:
      tags: [Enterprises]
      summary: Update current enterprise
      description: Updates the authenticated user's enterprise. Requires owner or admin role.
      operationId: updateCurrentEnterprise
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEnterpriseRequest"
      responses:
        "200":
          description: Enterprise updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enterprise"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/users/me:
    get:
      tags: [Users]
      summary: Get current user
      description: Returns the authenticated user's profile.
      operationId: getCurrentUser
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"

    patch:
      tags: [Users]
      summary: Update current user
      description: Updates the authenticated user's profile.
      operationId: updateCurrentUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/users/me/auth-providers/{provider}:
    delete:
      tags: [Users]
      summary: Disconnect an auth provider
      description: |
        Removes an OAuth identity provider link from the current user's account.
        Cannot disconnect the last auth method — the user must have another
        connected provider or a password set.
      operationId: disconnectAuthProvider
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
          description: Provider name to disconnect (e.g. google, github)
      responses:
        "204":
          description: Provider disconnected
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: array
              items:
                type: object
                additionalProperties: true
            request_id: { type: string }
...
    Enterprise:
      type: object
      required: [id, name, slug, plan, billing_email, created_at, updated_at]
      properties:
        id:
          type: string
          examples: [ent_abc123]
        name:
          type: string
        slug:
          type: string
          description: URL-friendly unique identifier
        plan:
          type: string
          enum: [free, pro, enterprise]
        personal:
          type: boolean
          default: false
          description: True for single-user personal accounts (e.g. Gmail sign-in)
        billing_email:
          type: string
          format: email
        sso_required:
          type: boolean
          default: false
        allowed_auth_providers:
          type: array
          items:
            type: string
          description: Restrict sign-in to these providers. Empty means all allowed.
        allowed_email_domains:
          type: array
          items:
            type: string
          description: Restrict sign-in to these email domains. Empty means all allowed.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    User:
      type: object
      description: |
        A user account. The id is an opaque surrogate key (usr_ + UUID). Email is
        the stable logical identity used to deduplicate across OAuth providers.
      required: [id, enterprise_id, email, display_name, role, status, auth_providers, has_password, created_at, updated_at]
      properties:
        id:
          type: string
          description: Opaque surrogate key — immutable, never changes
          examples: [usr_abc123]
        enterprise_id:
          type: string
        email:
          type: string
          format: email
          description: Unique email — the stable identity across OAuth providers
        display_name:
          type: string
        avatar_url:
          type: string
          format: uri
        role:
          type: string
          enum: [owner, admin, member]
        status:
          type: string
          enum: [active, pending_verification, invited, suspended]
        auth_providers:
          type: array
          items:
            $ref: "#/components/schemas/UserAuthProviderLink"
          description: Connected identity providers for this user
        has_password:
          type: boolean
          description: Whether the user has a password set for email/password login
        last_login_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserAuthProviderLink:
      type: object
      required: [provider, connected_at]
      properties:
        provider:
          type: string
          description: "Auth provider name (google, github, okta, saml, etc.)"
        connected_at:
          type: string
          format: date-time
...
    UpdateUserRequest:
      type: object
      properties:
        display_name:
          type: string
    UpdateEnterpriseRequest:
      type: object
      properties:
        name:
          type: string
        billing_email:
          type: string
          format: email
        sso_required:
          type: boolean
        allowed_auth_providers:
          type: array
          items:
            type: string
          description: Restrict sign-in to these providers. Empty array means all allowed.
        allowed_email_domains:
          type: array
          items:
            type: string
          description: Restrict sign-in to these email domains. Empty array means all allowed.
...